clap 98

+++ src/inf.base.format.vst3/inf.base.format.vst3/vst_controller.cpp
@@ -153,7 +153,7 @@ vst_controller::setParamNormalized(ParamID tag, ParamValue value)
   if(result != kResultOk) return result;
   update_state(tag);
   std::int32_t index = topology()->param_id_to_index.at(tag);
-  controller_param_changed(tag, format_normalized_to_base(topology(), true, index, static_cast<float>(value)));
+  controller_param_changed(tag, format_normalized_to_base(topology(), true, index, value));
   return kResultOk;
 }
